+++ server/src/main/java/nl/tudelft/oopp/group39/reservable/dto/FoodDto.java
@@ -26,7 +26,7 @@ public class FoodDto extends ReservableDto {
         String name,
         String description,
         Double price,
-        Integer building,
+        Long building,
         Set<ReservationAmountDto> reservations
     ) {
         super(price, building, initSet(reservations));
